requirements.txt
pathway
python-dotenv
google-generativeai
streamlit
requests
pandas
plotly
sqlalchemy
alembic
pydantic[email]
python-json-logger
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
aiohttp
numpy
scikit-learn